// src/gen/exa/reactive_component_pb_pb.ts
export class MessageDiff extends Message<MessageDiff> {
  /**
   * @generated from field: repeated exa.reactive_component_pb.FieldDiff field_diffs = 1;
   */
  fieldDiffs: FieldDiff[] = [];

  constructor(data?: PartialMessage<MessageDiff>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "exa.reactive_component_pb.MessageDiff";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "field_diffs", kind: "message", T: FieldDiff, repeated: true },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): MessageDiff {
    return new MessageDiff().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): MessageDiff {
    return new MessageDiff().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): MessageDiff {
    return new MessageDiff().fromJsonString(jsonString, options);
  }

  static equals(a: MessageDiff | PlainMessage<MessageDiff> | undefined, b: MessageDiff | PlainMessage<MessageDiff> | undefined): boolean {
    return proto3.util.equals(MessageDiff, a, b);
  }
}

/**
 * @generated from message exa.reactive_component_pb.FieldDiff
 */
export class FieldDiff extends Message<FieldDiff> {
  /**
   * @generated from field: uint32 field_number = 1;
   */
  fieldNumber = 0;

  /**
   * @generated from oneof exa.reactive_component_pb.FieldDiff.diff
   */
  diff: {
    /**
     * @generated from field: exa.reactive_component_pb.SingularValue update_singular = 2;
     */
    value: SingularValue;
    case: "updateSingular";
  } | {
    /**
     * @generated from field: exa.reactive_component_pb.RepeatedDiff update_repeated = 3;
     */
    value: RepeatedDiff;
    case: "updateRepeated";
  } | {
    /**
     * @generated from field: exa.reactive_component_pb.MapDiff update_map = 4;
     */
    value: MapDiff;
    case: "updateMap";
  } | {
    /**
     * @generated from field: bool clear = 5;
     */
    value: boolean;
    case: "clear";
  } | { case: undefined; value?: undefined } = { case: undefined };

  constructor(data?: PartialMessage<FieldDiff>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "exa.reactive_component_pb.FieldDiff";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "field_number", kind: "scalar", T: 13 /* ScalarType.UINT32 */ },
    { no: 2, name: "update_singular", kind: "message", T: SingularValue, oneof: "diff" },
    { no: 3, name: "update_repeated", kind: "message", T: RepeatedDiff, oneof: "diff" },
    { no: 4, name: "update_map", kind: "message", T: MapDiff, oneof: "diff" },
    { no: 5, name: "clear", kind: "scalar", T: 8 /* ScalarType.BOOL */, oneof: "diff" },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): FieldDiff {
    return new FieldDiff().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): FieldDiff {
    return new FieldDiff().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): FieldDiff {
    return new FieldDiff().fromJsonString(jsonString, options);
  }

  static equals(a: FieldDiff | PlainMessage<FieldDiff> | undefined, b: FieldDiff | PlainMessage<FieldDiff> | undefined): boolean {
    return proto3.util.equals(FieldDiff, a, b);
  }
}

/**
 * @generated from message exa.reactive_component_pb.RepeatedDiff
 */
export class RepeatedDiff extends Message<RepeatedDiff> {
  /**
   * @generated from field: uint32 new_length = 1;
   */
  newLength = 0;

  /**
   * @generated from field: repeated exa.reactive_component_pb.SingularValue update_values = 2;
   */
  updateValues: SingularValue[] = [];

  /**
   * @generated from field: repeated uint32 update_indices = 3;
   */
  updateIndices: number[] = [];

  constructor(data?: PartialMessage<RepeatedDiff>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "exa.reactive_component_pb.RepeatedDiff";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "new_length", kind: "scalar", T: 13 /* ScalarType.UINT32 */ },
    { no: 2, name: "update_values", kind: "message", T: SingularValue, repeated: true },
    { no: 3, name: "update_indices", kind: "scalar", T: 13 /* ScalarType.UINT32 */, repeated: true },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): RepeatedDiff {
    return new RepeatedDiff().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): RepeatedDiff {
    return new RepeatedDiff().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): RepeatedDiff {
    return new RepeatedDiff().fromJsonString(jsonString, options);
  }

  static equals(a: RepeatedDiff | PlainMessage<RepeatedDiff> | undefined, b: RepeatedDiff | PlainMessage<RepeatedDiff> | undefined): boolean {
    return proto3.util.equals(RepeatedDiff, a, b);
  }
}

/**
 * @generated from message exa.reactive_component_pb.MapDiff
 */
export class MapDiff extends Message<MapDiff> {
  /**
   * @generated from field: repeated exa.reactive_component_pb.MapKeyDiff map_key_diffs = 1;
   */
  mapKeyDiffs: MapKeyDiff[] = [];

  constructor(data?: PartialMessage<MapDiff>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "exa.reactive_component_pb.MapDiff";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "map_key_diffs", kind: "message", T: MapKeyDiff, repeated: true },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): MapDiff {
    return new MapDiff().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): MapDiff {
    return new MapDiff().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): MapDiff {
    return new MapDiff().fromJsonString(jsonString, options);
  }

  static equals(a: MapDiff | PlainMessage<MapDiff> | undefined, b: MapDiff | PlainMessage<MapDiff> | undefined): boolean {
    return proto3.util.equals(MapDiff, a, b);
  }
}

/**
 * @generated from message exa.reactive_component_pb.MapKeyDiff
 */
export class MapKeyDiff extends Message<MapKeyDiff> {
  /**
   * @generated from field: exa.reactive_component_pb.SingularValue map_key = 1;
   */
  mapKey?: SingularValue;

  /**
   * @generated from oneof exa.reactive_component_pb.MapKeyDiff.diff
   */
  diff: {
    /**
     * @generated from field: exa.reactive_component_pb.SingularValue update_singular = 2;
     */
    value: SingularValue;
    case: "updateSingular";
  } | {
    /**
     * @generated from field: bool clear = 3;
     */
    value: boolean;
    case: "clear";
  } | { case: undefined; value?: undefined } = { case: undefined };

  constructor(data?: PartialMessage<MapKeyDiff>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "exa.reactive_component_pb.MapKeyDiff";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "map_key", kind: "message", T: SingularValue },
    { no: 2, name: "update_singular", kind: "message", T: SingularValue, oneof: "diff" },
    { no: 3, name: "clear", kind: "scalar", T: 8 /* ScalarType.BOOL */, oneof: "diff" },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): MapKeyDiff {
    return new MapKeyDiff().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): MapKeyDiff {
    return new MapKeyDiff().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): MapKeyDiff {
    return new MapKeyDiff().fromJsonString(jsonString, options);
  }

  static equals(a: MapKeyDiff | PlainMessage<MapKeyDiff> | undefined, b: MapKeyDiff | PlainMessage<MapKeyDiff> | undefined): boolean {
    return proto3.util.equals(MapKeyDiff, a, b);
  }
}